backend/users/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-dotenv==1.1.1
argon2-cffi==25.1.0
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.5.1